本人認証と二要素認証クラスの別け
